requirements.txt
streamlit
nltk
sentence-transformers
chromadb
scikit-learn
requests
matplotlib
seaborn
torch
pandas
numpy
transformers
rouge
tqdm
openai
mcp
groq
fastapi
uvicorn[standard]
pydantic
aiohttp
email-validator
python-dotenv
textblob
langchain
langchain-community